requirements.txt
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
snowflake-connector-python>=3.10.0
